[feat] discussions 로그인 작성자 권한처리

+++ discussions/views.py
@@ -1,9 +1,8 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import DiscussionPost, DiscussionComment
-from django.contrib.auth import get_user_model # 임시
-# from django.contrib.auth.decorators import login_required
-# from django.http import HttpResponseForbidden
-User = get_user_model() # 임시
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
+
 
 # Create your views here.
 def post_list(request):
@@ -26,7 +25,7 @@ def post_detail(request, post_id):
         'comments':comments
     })
 
-# @login_required
+@login_required
 def post_create(request):
     if request.method == 'POST':
         title = request.POST.get('title')
@@ -35,7 +34,7 @@ def post_create(request):
         image = request.FILES.get('image')
 
         DiscussionPost.objects.create(
-            author=User.objects.first(), # 임시 -> author=request.user
+            author=request.user,
             title=title,
             content=content,
             link=link,
@@ -45,7 +44,7 @@ def post_create(request):
     
     return render(request,'discussions/post_form.html')
 
-# @login_required
+@login_required
 def comment_create(request,post_id):
     post = get_object_or_404(DiscussionPost,id=post_id)
     if request.method == 'POST':
@@ -53,17 +52,17 @@ def comment_create(request,post_id):
 
         DiscussionComment.objects.create(
             post=post,
-            author=User.objects.first(), # 임시 -> author=request.user
+            author=request.user,
             content=content
         )
     return redirect('discussions:post_detail',post_id=post_id)
 
-# @login_required
+@login_required
 def post_edit(request,post_id):
     post = get_object_or_404(DiscussionPost,id=post_id)
 
-    # if post.author != request.user:
-    #    return HttpResponseForbidden("수정 권한이 없습니다.")
+    if post.author != request.user:
+        return HttpResponseForbidden("수정 권한이 없습니다.")
 
     if (request.method == 'POST'):
         post.title=request.POST.get('title')
@@ -75,36 +74,36 @@ def post_edit(request,post_id):
         return redirect('discussions:post_detail',post_id=post.id)
     return render(request,'discussions/post_form.html',{'post':post})
 
-# @login_required
+@login_required
 def post_delete(request,post_id):
     post = get_object_or_404(DiscussionPost,id=post_id)
 
-    # if post.author != request.user:
-    #    return HttpResponseForbidden("삭제 권한이 없습니다.")
+    if post.author != request.user:
+       return HttpResponseForbidden("삭제 권한이 없습니다.")
 
     if request.method == 'POST':
         post.delete()
     return redirect('discussions:post_list')
 
-# @login_required
+@login_required
 def comment_delete(request,comment_id):
     comment = get_object_or_404(DiscussionComment,id=comment_id)
     post_id = comment.post.id
 
-    # if comment.author != User.objects.first():
-    #    return HttpResponseForbidden("삭제 권한이 없습니다.")
+    if comment.author != request.user:
+       return HttpResponseForbidden("삭제 권한이 없습니다.")
 
     if request.method == 'POST':
         comment.delete()
     return redirect('discussions:post_detail',post_id=post_id)
 
-# @login_required
+@login_required
 def comment_edit(request,comment_id):
     comment = get_object_or_404(DiscussionComment,id=comment_id)
     post_id = comment.post.id
 
-    # if comment.author != User.objects.first():
-    #    return HttpResponseForbidden("수정 권한이 없습니다.")
+    if comment.author != request.user:
+       return HttpResponseForbidden("수정 권한이 없습니다.")
 
     if request.method == 'POST':
         comment.content = request.POST.get('content')
